+++ EssentialFeed/EssentialFeediOS/FeedUI/Composers/FeedLoaderPresentationAdapter.swift
@@ -1,0 +1,31 @@
+//
+//  FeedLoaderPresentationAdapter.swift
+//  EssentialFeed
+//
+//  Created by Israel Marcos Alvarez Mesa on 24/1/26.
+//
+
+import EssentialFeed
+
+final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
+    private let feedLoader: FeedLoader
+    var presenter: FeedPresenter?
+
+    init(feedLoader: FeedLoader) {
+        self.feedLoader = feedLoader
+    }
+
+    func didRequestFeedRefresh() {
+        presenter?.didStartLoadingFeed()
+
+        feedLoader.load { [weak self] result in
+            switch result {
+            case let .success(feed):
+                self?.presenter?.didFinishLoadingFeed(with: feed)
+
+            case let .failure(error):
+                self?.presenter?.didFinishLoadingFeed(with: error)
+            }
+        }
+    }
+}
